requirements.txt
Flask
google-generativeai
numpy
sentence-transformers
torch
PyMuPDF>=1.23.0
python-docx>=0.8.11
nltk>=3.8.1
python-pptx>=0.6.21
langchain
faiss-cpu
imageio[ffmpeg]
transformers
Pillow